卸载对象报错处理 #2517

+++ services/service-metadata-objects/src/lookup/LookupActionHandler.ts
@@ -18,6 +18,9 @@ export class LookupActionHandler {
 
     async deleteAll(objectConfig){
         try {
+            if (!objectConfig) {
+                return;
+            }
             await this.deleteAllLookups(objectConfig);
             return true
         } catch (error) {
